+++ packages/extension/test/integration/v2/background_spec.js
@@ -848,6 +848,9 @@ describe('app/background', () => {
       beforeEach(() => {
         sinon.stub(browser.windows, 'getCurrent').withArgs({ populate: true }).resolves({ id: '10', tabs: [{ id: '1' }, { id: '2' }, { id: '3' }] })
         sinon.stub(browser.tabs, 'remove').withArgs(['1', '2', '3']).resolves()
+        sinon.stub(browser.tabs, 'create').withArgs({ url: 'about:blank', active: false }).resolves({
+          id: 'new-tab',
+        })
       })
 
       it('closes the tabs in the current browser window', function (done) {
@@ -857,11 +860,35 @@ describe('app/background', () => {
 
           expect(browser.windows.getCurrent).to.be.called
           expect(browser.tabs.remove).to.be.called
+          expect(browser.tabs.create).not.to.be.called
 
           done()
         })
 
         this.server.emit('automation:request', 123, 'reset:browser:tabs:for:next:test')
+      })
+
+      // @see https://github.com/cypress-io/cypress/issues/29172
+      describe('firefox 124 and up', () => {
+        beforeEach(() => {
+          global.window.navigator = {
+            userAgent: 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10.15; rv:124.0) Gecko/20100101 Firefox/124.0',
+          }
+        })
+
+        it('creates a new "about:blank" tab and closes the other tabs in the current browser window', function (done) {
+          this.socket.on('automation:response', (id, obj) => {
+            expect(id).to.eq(123)
+            expect(obj.response).to.be.undefined
+
+            expect(browser.windows.getCurrent).to.be.called
+            expect(browser.tabs.remove).to.be.calledWith(['1', '2', '3'])
+            expect(browser.tabs.create).to.be.calledWith({ url: 'about:blank', active: false })
+            done()
+          })
+
+          this.server.emit('automation:request', 123, 'reset:browser:tabs:for:next:test')
+        })
       })
     })
   })